refactor: move mutation functions to utils
