+++ src/cachers/index.js
@@ -1,8 +1,6 @@
-import pick from 'lodash/pick.js';
 import defu from 'defu';
 import * as utils from './utils.js';
 import nuxtConfig from '../../nuxt.config.js';
-import { langMapValueForLocale } from '../plugins/europeana/utils.js';
 import localeCodes from '../plugins/i18n/codes.js';
 
 const CACHE_KEY_PREFIX = '@europeana:portal.js';
@@ -28,43 +26,6 @@ const cacherModule = (cacherName) => {
   return import(`./${cacherPath}.js`);
 };
 
-const dailyOffset = (setSize, subsetSize) => {
-  const millisecondsPerDay = (1000 * 60 * 60 * 24);
-  const unixDay = Math.floor(Date.now() / millisecondsPerDay);
-  const offset = (unixDay * subsetSize) % setSize;
-  return (offset + subsetSize <= setSize) ? offset : (setSize - subsetSize);
-};
-
-const daily = (set, subsetSize) => {
-  if (!Array.isArray(set)) {
-    return set;
-  }
-  const offset = dailyOffset(set.length, subsetSize);
-  return set.slice(offset, offset + subsetSize);
-};
-
-const localiseOne = (item, fields, locale) => {
-  const localised = { ...item };
-
-  for (const field of [].concat(fields)) {
-    if (localised[field]) {
-      localised[field] = langMapValueForLocale(localised[field], locale).values[0];
-    }
-  }
-
-  return localised;
-};
-
-const localise = (data, fields, locale) => {
-  if (Array.isArray(data)) {
-    return data.map(item => localiseOne(item, fields, locale));
-  } else if (typeof data === 'object') {
-    return localiseOne(data, fields, locale);
-  } else {
-    return data;
-  }
-};
-
 const cacheKey = (cacherName, locale) => {
   let key = CACHE_KEY_PREFIX;
 
@@ -76,27 +37,24 @@ const cacheKey = (cacherName, locale) => {
   return key;
 };
 
-// TODO: move conditional mutations to own script files?
 const runSetCacher = async(cacherName) => {
   console.log(cacherName);
+
   const cacher = await cacherModule(cacherName);
   let data = await cacher.data(runtimeConfig);
 
-  if (cacher.DAILY && Array.isArray(data)) {
-    data = daily(data, cacher.DAILY);
+  if (cacher.DAILY) {
+    data = utils.daily(data, cacher.DAILY);
   }
 
   if (cacher.PICK) {
-    if (Array.isArray(data)) {
-      data = data.map(item => pick(item, cacher.PICK));
-    } else if (typeof data === 'object') {
-      data = pick(data, cacher.PICK);
-    }
+    data = utils.pick(data, cacher.PICK);
   }
 
   if (cacher.LOCALISE) {
     for (const locale of localeCodes) {
-      await writeCacheKey(cacheKey(cacherName, locale), localise(data, cacher.LOCALISE, locale));
+      const localised = utils.localise(data, cacher.LOCALISE, locale);
+      await writeCacheKey(cacheKey(cacherName, locale), localised);
     }
   } else {
     await writeCacheKey(cacheKey(cacherName), data);
