+++ src/cachers/utils.js
@@ -1,6 +1,8 @@
 import axios from 'axios';
 import redis from 'redis';
+import _pick from 'lodash/pick.js';
 import { promisify } from 'util';
+import { langMapValueForLocale } from '../plugins/europeana/utils.js';
 
 const redisConfig = (config = {}) => {
   const redisOptions = {
@@ -53,8 +55,56 @@ const errorMessage = (error) => {
   return message;
 };
 
+const dailyOffset = (setSize, subsetSize) => {
+  const millisecondsPerDay = (1000 * 60 * 60 * 24);
+  const unixDay = Math.floor(Date.now() / millisecondsPerDay);
+  const offset = (unixDay * subsetSize) % setSize;
+  return (offset + subsetSize <= setSize) ? offset : (setSize - subsetSize);
+};
+
+const daily = (set, subsetSize) => {
+  if (!Array.isArray(set)) {
+    return set;
+  }
+  const offset = dailyOffset(set.length, subsetSize);
+  return set.slice(offset, offset + subsetSize);
+};
+
+const localiseOne = (item, fields, locale) => {
+  const localised = { ...item };
+
+  for (const field of [].concat(fields)) {
+    if (localised[field]) {
+      localised[field] = langMapValueForLocale(localised[field], locale).values[0];
+    }
+  }
+
+  return localised;
+};
+
+const localise = (data, fields, locale) => {
+  return mutateObjects(data, object => localiseOne(object, fields, locale));
+};
+
+const pick = (data, fields) => {
+  return mutateObjects(data, object => _pick(object, fields));
+};
+
+const mutateObjects = (data, mutator) => {
+  if (Array.isArray(data)) {
+    return data.map(mutator);
+  } else if (typeof data === 'object') {
+    return mutator(data);
+  } else {
+    return data;
+  }
+};
+
 export {
   createEuropeanaApiClient,
   createRedisClient,
-  errorMessage
+  errorMessage,
+  daily,
+  localise,
+  pick
 };
